requirements.txt
aiohttp>=3.11
PyYAML>=6.0
pandas>=2.2
numpy>=1.24
ta-lib==0.5.1
